+++ src/main/java/com/verdantartifice/primalmagic/common/items/ItemsPM.java
@@ -326,6 +326,7 @@ public class ItemsPM {
     public static final RegistryObject<BlockItem> HEXIUM_BLOCK = ITEMS.register("hexium_block", () -> new BlockItem(BlocksPM.HEXIUM_BLOCK.get(), new Item.Properties().tab(PrimalMagic.ITEM_GROUP)));
     public static final RegistryObject<BlockItem> HALLOWSTEEL_BLOCK = ITEMS.register("hallowsteel_block", () -> new BlockItem(BlocksPM.HALLOWSTEEL_BLOCK.get(), new Item.Properties().tab(PrimalMagic.ITEM_GROUP)));
     public static final RegistryObject<BurnableBlockItem> IGNYX_BLOCK = ITEMS.register("ignyx_block", () -> new BurnableBlockItem(BlocksPM.IGNYX_BLOCK.get(), 128000, new Item.Properties().tab(PrimalMagic.ITEM_GROUP)));
+    public static final RegistryObject<BlockItem> SALT_BLOCK = ITEMS.register("salt_block", () -> new BlockItem(BlocksPM.SALT_BLOCK.get(), new Item.Properties().tab(PrimalMagic.ITEM_GROUP)));
     
     // Register salted food items
     public static final RegistryObject<Item> SALTED_BAKED_POTATO = ITEMS.register("salted_baked_potato", () -> new Item(new Item.Properties().tab(PrimalMagic.ITEM_GROUP).food(new FoodProperties.Builder().nutrition(6).saturationMod(0.72F).build())));