+++ changelog.md
@@ -11,6 +11,7 @@ v0.0.18-alpha
 * The Ritual Altar now requires two empty blocks above it to start a ritual
 * Fix clumping of mana cost widgets on arcane workbench screen
 * Increased affinities of glowstone blocks
+* Add storage block for Refined Salt
 
 v0.0.17-alpha
 -------------